--- project/src/store/favorite-offers/action.ts
@@ -1,4 +1,0 @@
-import {createAction} from '@reduxjs/toolkit';
-import {Offers} from 'types/offer';
-
-export const setFavoriteOffers = createAction<Offers>('favorite-offers/setFavoriteOffers');
