+++ project/src/store/reducer.ts
@@ -1,6 +1,6 @@
 import {combineReducers} from '@reduxjs/toolkit';
 import {cityReducer} from 'store/city/city-reducer';
-import {favoriteOffersReducer} from 'store/favorite-offers/reducer';
+import {favoriteOffersReducer} from 'store/favorite-offers/favorite-offers-reducer';
 import {offerReducer} from 'store/offer/reducer';
 import {offersReducer} from 'store/offers/reducer';
 import {StoreNamespace} from 'settings/store-namespace';
@@ -8,7 +8,7 @@ import {userReducer} from 'store/user/reducer';
 
 export const reducer = combineReducers({
   [StoreNamespace.City]: cityReducer.reducer,
-  favoriteOffersReducer,
+  [StoreNamespace.FavoriteOffers]: favoriteOffersReducer.reducer,
   offerReducer,
   offersReducer,
   userReducer,
